Fix duplicate articles in archive.
During the initial article import, many of the articles had the same
created_at.  It looks like the sort wasn't deterministic when only
sorting by created_at causing some articles to be duplicated when
paginating.  This commit adds a secondary sort by article title so that
it will always sort them in the same way.

Closes #359