Make asset path match `assets/articles/YYYY/MM/DD`
I also added a way for the header to be named 'header-X' where X
is (1..∞) based on how many articles a found published already on a given day.

My concern is there is not a way to make sure the header is matched to the right article.

for example:

1. create 2 articles, article-1 and article-2, to be published on the same day, in that order
2. add a header to article-2 before you add one to article-1
3. now article-1 will map to header-2 and article-2 will map to header-1

That is a bit confusing (even typing it was confusing) so I need to figure out how to
keep the order correct