Check passwords on user sign-up against DB of known leaks
I was going through a bunch of accounts on [HaveIBeenPwned][0] and I
noticed the [Passwords][1] tab for the first time.

That led me to reading this [post][2] about checking against this
password DB.

This PR specifically implements the registration use case [mentioned here][3].

There are still several technical issue to resolve, but I wanted to
open a PR to start getting feedback before continuing.

My open questions are:

1. Do we want to do this?
2. Would we want to implement the other Use Cases mentioned?
3. If so, what is the best way to store this giant password DB (It's 5.7 GB *compressed*)
4. A faster way to do the DB lookup (probably don't want to use ActiveRecord for this)

[0]:https://haveibeenpwned.com/
[1]:https://haveibeenpwned.com/Passwords
[2]:https://www.troyhunt.com/introducing-306-million-freely-downloadable-pwned-passwords/
[3]:https://www.troyhunt.com/introducing-306-million-freely-downloadable-pwned-passwords/#usecase1registration