[TCE] normalize variable images
When different languages have different images, we swap
between them by using a convention with the image paths.

e.g.

english version:
https://cloudfront.crimethinc.com/tce/images/revolt-large.jpg

spanish version
https://cloudfront.crimethinc.com/tce/images/rebelarte-large.jpg

This requires the same image variations to be uploaded for every language,
otherwise one language will look fine, the other will have an image missing.

I noticed that for espanol we have:
https://cloudfront.crimethinc.com/tce/images/rebelarte-large@2x.jpg

But we do not have the equivalent file in the english version.