Add spec to lock in current Pages behavior
This PR adds a system spec that verifies the current behavior of the
`Page`s we have in production

Current Behavior
---------------------
Currently in production we have the following pages
```ruby
```

Also, currently, if a page doesn't exist we currently re-direct to the
homepage instead of throwing a `404`

Why?
-------
We want to deprecate the `Page` feature (it isn't really used) and get
rid of the catchall redirect so that we start showing `404`s again.

The main two reasons for showing `404`s again are:
1. we are pretty stable now, and would be nice to fix broken links in the wild with redirects
2. this should make integrating with `ActiveStorage` easier as there
   is currently some funky behavior with the catchall route

This PR Does Not
----------------
Note that this PR doesn't change any of the current behavior (leaves
the catchall route in place)

This PR just locks in a spec for the current content so that when the
route file is changed in a future PR, we don't regress any of these pages