[i18n] Make all of the links on the language landing page work
This constructs a in-memory lookup table for translating locales that
come in via URL params into the canonical URL slug we have in the DB.

This allows both english-ified names, local names, unicode names,
etc. to all vbe used in the URL just work.

For example, both of these work:
- localhost:3000/languages/ελληνικά
- localhost:3000/languages/greek

I don't really like how many layers of redirects we have already so
didn't want to add another

The specs make sure that all of the links in `/languages` `200` (prior
to this change some of them would `500`)