Add  JSON feed template for the tools controller
We started getting `500`s for URLs that look like this:

```
https://crimethinc.com/zines.json
```

I went back through the git history to August of
2021 (ca65aedc4e399715b04114bafc9129a5b29d747a) to see if this was a
recent regression. It seems that we **never had JSON feed support for tools**
in the first place.

To add JSON support for tools, we will need to:

1. Add a json index template
2. modify the tool controller's namespace method to support json

This commit handles number 1