Add `rescue` for ActiveStorage image uploads
in development/staging we often download production data and load
it into our local database to facilitate development and
debugging. This causes an error with image and resources that are
uploaded via ActiveStorage. We get the below error whenever trying
to generate an image url:

```
 ActionView::Template::Error (undefined method `signed_id' for nil:NilClass):
```

This is because the development/staging ActiveStorage tables don't
(and cannot) match up with the actual ActiveStorage storage used
in these environments (e.g. the staging env points to aa different
AWS instance than the production env)

This is a bit of a stop-gap until we figure out a better way of
handling this issue.

I re-throw the error in production to prevent us from swallowing legit
errors