use `controller_path` for tool namespacing
When a request for an HTML page comes in, `request.path` looks like
this:

```ruby
[2] pry(#<StickersController>)> request.path
=> "/stickers"
```

The existing code then strips off the leading `/` and just returns
the tool name (`zines` in this example)

An issues occurs when we get a JSON request. That is because
`request.path` will look like this:

```ruby
[1] pry(#<StickersController>)> request.path
=> "stickers.json"
```

The existing code will then incorrectly return `stickers.json` as the tool
namespace, which cause downstream issues

replacing the code with `controller_path` will return the controllers
name the same regardless of format, so this will hopefully have fewer
issues.

Note: Because we have this `tools` abstraction, I am not 100% sure
that the `controller_path` will always be the same as the expected
tool. I couldn't find an instance of this currently, but it doesn't
seem impossible