Add json feed for zines (#2369)
* Add  JSON feed template for the tools controller

We started getting `500`s for URLs that look like this:

```
https://crimethinc.com/zines.json
```

I went back through the git history to August of
2021 (ca65aedc4e399715b04114bafc9129a5b29d747a) to see if this was a
recent regression. It seems that we **never had JSON feed support for tools**
in the first place.

To add JSON support for tools, we will need to:

1. Add a json index template
2. modify the tool controller's namespace method to support json

This commit handles number 1

* Add `rescue` for ActiveStorage image uploads

in development/staging we often download production data and load
it into our local database to facilitate development and
debugging. This causes an error with image and resources that are
uploaded via ActiveStorage. We get the below error whenever trying
to generate an image url:

```
 ActionView::Template::Error (undefined method `signed_id' for nil:NilClass):
```

This is because the development/staging ActiveStorage tables don't
(and cannot) match up with the actual ActiveStorage storage used
in these environments (e.g. the staging env points to aa different
AWS instance than the production env)

This is a bit of a stop-gap until we figure out a better way of
handling this issue.

I re-throw the error in production to prevent us from swallowing legit
errors

* use `controller_path` for tool namespacing

When a request for an HTML page comes in, `request.path` looks like
this:

```ruby
[2] pry(#<StickersController>)> request.path
=> "/stickers"
```

The existing code then strips off the leading `/` and just returns
the tool name (`zines` in this example)

An issues occurs when we get a JSON request. That is because
`request.path` will look like this:

```ruby
[1] pry(#<StickersController>)> request.path
=> "stickers.json"
```

The existing code will then incorrectly return `stickers.json` as the tool
namespace, which cause downstream issues

replacing the code with `controller_path` will return the controllers
name the same regardless of format, so this will hopefully have fewer
issues.

Note: Because we have this `tools` abstraction, I am not 100% sure
that the `controller_path` will always be the same as the expected
tool. I couldn't find an instance of this currently, but it doesn't
seem impossible

* Update app/views/2017/tools/index.json.jbuilder

Co-authored-by: Shane Becker <veganstraightedge@gmail.com>