Validate that sitemap.xml is not redirected
I don't know why, but sitemap.xml is not redirected by our middleware. I
thought I'd ahve to add `|| @req.path == '/sitemap.xml'` to the
`exit_early?` method, but it's not the case.

To validate I add the `public/sitemap.xml` file and I validate the I got
the right content when I query `/sitemap.xml` and it works. The only
things to keep in mind is since I create a dummy `public/sitemap.xml`
file, we need to regenerate one every time we deploy the application.