[BUG] Update locale.articles_count when an article is destroyed
We have an `after_save` hook that will look up an article's locale,
and update the `Locale.articles_count` value for the article's
associated locale.

This value is used in rendering the [languages page][0] on the site.

We encountered a situation where we had 0 Galician articles in the
database, yet the website was displaying a `1` for Galician on the
site.

After looking into the code we realized that our `after_save` hook
only covered the use case of an article being created or updated, not
destroyed. There was also a bug with the `update` case.

The two bugs are:

- If an article is destroyed, we never trigger the callback. This
  doesn't have a huge effect on locale's with a lot of articles, as
  evneually the value is updated. However, locales with infrequent
  articles can get into a bad state and stay there for awhile

- If an article has its locale changed, the original locale is never
  updated, only the new locale it has been changed to.

[0]:https://crimethinc.com/languages