Migrate live blog polling to Stimulus
I migrated the live blog polling logic to stimulus. In the original
code there was a queueing mechanism set up, I think with the hope that
eventually we might consume messages off of the queue and actually
update the page's content without a refresh

That never happened and all it actually does is pop up a banner that
alerts the user to refresh their page to see new posts. That is the
featureset I migrated to stimulus.

The live updating can more easily be accomplished with hotwire now
that that is an option, so perhaps we will revist this feature going
forward. In the meantime the live blog functionality as it stands with
this PR is:

- If an article is a collection, this live_blog controller is loaded
  with the UI in a hidden state
- On page load, the current UTC time is stored in the
  controler (`pageLoadedAt`)
- in the `initialize()` hook, the poller is started with a 1 minute
  interval
- I modified the collection endpoint to accept the parent articles
  slug and the `pageLoadedAt` utc value
- The endpoint queries for any `article.collection_posts` published
  after `pageLoadedAt`
- when the controller receives the response, it just updates the
  `updateCount` value with the length of the returned posts array
- if the value changes and is greater than 0, the onchange callback
  updates the banners text and makes it visible

A note on the changes in the collection endpoint: I left the original
codepath in place just in case someone else was consuming this
endpoint, but I think that original path is now dead code. Maybe when
we revisit this feature we can clean that up