migrate support.js to importmaps
I do think the support page slider is a good candidate for a stimulus
controller.

That said, I do not know how to set stripe up for testing so I decided
to not modify any of that javascript. Instead I am using this as an
example of how to selectivley load some custom JS on specific pages
with import maps

1. I moved the file from the `app/assets/javascript` directory to the
   `app/javascript` directory as this is where JS will be placed for
   import maps
2. I removed the file from the asset pipeline manifest (I think this
   is correct, but I need to double check that)
3. I used the `content_for :head` to load the JS selectivley in
   `/support` views

I checked that the JS file does indeed load on the `/support` change
with these changes (at least in development)