Remove twitter deps (#3312)
This PR also add `ruby-lsp` as a dev deps since it's really easier for
me to have it inside the project and not system-wide.

The main goal of this PR is to remove the dependency to the `twitter`
gem. I also remove the `Twitter` model since it was the class using the
`TwitterClient`. I was wondering if you want to also remove the
`RandomTools` class? I didn't remove it for now since it wasn't coupled
to the `Twitter` model, but we don't use it anywhere else, so unless you
want to create the same kind of bot, but for mastodon, I'd also remove
the `RandomTools` class.

---------

Co-authored-by: Shane Becker <veganstraightedge@gmail.com>