Add auto-merge for dependabot PR except when it's a major upgrade (#3511)
I give up, there's to many dependencies updates. There'll always be a
risk of supply chain attack, even if we review dependencies updates
manually. The key will probably be to keep the number of dependencies as
low as possible, but I wanna try auto-merge to see if it reduces the
amount of work.