Fix social share preview for Posters (#4707)
# Before

`Poster#meta_image` was returning an `ActiveStorage` object `to_s`,
instead of a URL.
The view is expecting a plain string URL, which it'll use in a `<meta>`
tag.

```html
<meta name="twitter:image" content="#<ActiveStorage::Attached::One:0x00007f2a3e2b8468>" property="og:image">
```

# After

`Poster#meta_image` returns the `url_for` of an `ActiveStorage`
attachment on `Poster`.
It prefers `#image_front_color_image`, then
`image_front_black_and_white_image`.
If neither is `.attached?`, it returns `nil`, allowing the `meta_image`
view to fall back to the site wide default image (for now).

```html
<meta name="twitter:image" content="https://crimethinc.com/rails/active_storage/blobs/proxy/eyJfcmFpbHMiOnsiZGF0YSI6MjE1LCJwdXIiOiJibG9iX2lkIn19--8cdcc186c36ac03e15371c5a6cbc1181765fe9e6/where-are-we-going_front_black_and_white.gif" property="og:image">
```