Add locale subdomains: ceb es-419 fo fr-qu li lt ms mt (#4723)
We have something translated to all of these languages. Either the site,
To Change Everything, tools (books, zines, etc), or most commonly
articles.

The goal is to have a subdomain with a site translation for any language
that we have content for.

More work is needed to fully support these as subdomains (DNS, YAML
files, etc).
This is just one of the first steps. 


| code   | Language/dialect           |
| ------ | -------------------------- |
| fr-qu  | French Canadian/Quebecois  |
| es-419 | Spanish (Latin America)    |
| ceb    | Cebuano                    |
| fo     | Faroese                    |
| li     | Lithuanian, redirect to LT |
| lt     | Lithuanian, actual         |
| ms     | Malay                      |
| mt     | Maltese                    |