Add # anchor links to article content headings (#4756)
# Context

Our headings already get transliterated (slug safe) ID attributes on the
heading tag:

```markdown
# Intro
```

```html
<h1 id='intro'>Intro</h1>
```

The are, of course, addressable in a URL by adding `#intro` (in this
example) to the URL.
For a simple one word example, that's easy enough.
But for most headings, one has to either manually sluggify the heading
content into an anchor or inspect the HTML to copy it from the `h1@id`
value. Neither are friendly.

We use a Markdown parser called
[kramdown](https://kramdown.gettalong.org).
By setting `header_links: true` when we render Markdown `.to_html`, we
tell kramdown to generate links in the heading with the right anchor
`href` value for us.

Sadly, kramdown doesn't seem to have a way to specify what the link
content should be or where to position it. It only creates an empty
`<a>` tag before the heading content.

That's what the CSS is for, to give it content `#` and position it out
in the margin next to the article content.

I've opened an issue in the kramdown gem repo, asking for options to
specify these. All things being equal, I'd rather render the link text
server side with the template and place it after the heading content.
(Which I tried with CSS, but it gets weird if the heading wraps to
multiple lines, which is common in our articles.)

---

# Before

<img width="1944" height="614" alt="Screenshot 2025-08-01 at 6 40 05 PM"
src="https://github.com/user-attachments/assets/0c71a92a-7f51-418b-9e51-f3c7aa30bfb1"
/>

# Hover on a heading

Shows a `#` link to the left of the heading.

<img width="1944" height="612" alt="Screenshot 2025-08-01 at 6 40 13 PM"
src="https://github.com/user-attachments/assets/1be67c1a-2a44-4110-b36e-d822b075748a"
/>

# Clicked link

Scrolls the page to that heading, changes the URL to `URL#heading-id`

<img width="1750" height="336" alt="Screenshot 2025-08-01 at 6 40 49 PM"
src="https://github.com/user-attachments/assets/639dab5a-cff2-4248-bab6-825dbea84f57"
/>