TCE cebuano web version
This creates a web version of the cebuano TCE pdf.

It also removes links to a tagalog version of the PDF from the TCE
menus because we do not actually have a tagalog translation yet.

The website mentions that the PDF text is in cebuano, while the video
is in tagalog, which is why we thought we had a tagalog pdf

relates to: https://github.com/crimethinc/website/issues/574
closes: https://github.com/crimethinc/website/issues/1400
closes: https://github.com/crimethinc/website/issues/1385