Create web version of the Malaysian TCE pdf (#4758)
## This change

Malay comes in many dialects, so I am *assuming* "standard malay" is
used in the pdf.

The generator and guide appears to be working well, though the
`video_url` in the generated yaml is in the wrong section. The
generator should be updated to correct that.

## Resources used

-
https://github.com/crimethinc/website/blob/main/docs/to-change-everything-guide.md
- https://en.wikipedia.org/wiki/Malaysian_Malay

# What are the relevant GitHub issues?

relates to: https://github.com/crimethinc/website/issues/575
closes: https://github.com/crimethinc/website/issues/1392