[TCE] update CSS for double column TOC
While working on the Italian TCE, I had to fix the table of contents
(TOC) for the phone/tablet device sizes.

After checking, it looks like this menu is misaligned for most
languages.

I have updated the flexbox CSS to:

- set the header for the menu to full width

- with the header a full width, the two columns now wrap to the next
  flexbox row.

- I set each of the columns to take up 47.5% of the row (this leaves
  5% leftover on the row)

- I set the `justify-content: space-between' so that the 5% leftover
becomes a gap between the columns.

The `text-align: start' and `text-indent: 10px hanging' achieve a
similar visual effect as the previous `text-indent: -5vw', without
breaking out of the box boundaries.