[TCE] Initial Srpskohrvatski
This initial pass on converting the Srpskohrvatski PDF to the web
version slots in all of the translations available in the PDF.

There are things that need to be translated that are not in the PDF,
such as menu items (e.g. "Download PDF"), certain headers (e.g. "The
secret is to begin"), image alt text, etc.

Normally I throw these things into google translate, as the original
translators are unavailable.

I have done that here also, with an extra complication I wanted to
call out. The text is is serbocroation, which isn't an option in
google translate. I went with "bosnian" in google translate as that
was the most common "autodetected" language when I pasted in different
sections of the PDF translations.

---

resolves: https://github.com/crimethinc/website/issues/1397
related to: https://github.com/crimethinc/website/issues/575